EVSS document upload refactor, no functional change (#6463)
* move code to module

* rearranging, no functional change

* fix refactor style

* coments only

* use fewer doubles in specs

* style
