+++ app/models/supporting_evidence_attachment.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 # Files uploaded as part of a form526 submission that will be sent to EVSS upon form submission.
+# inherits from ApplicationRecord
 class SupportingEvidenceAttachment < FormAttachment
   ATTACHMENT_UPLOADER_CLASS = SupportingEvidenceAttachmentUploader
 end
