+++ app/uploaders/supporting_evidence_attachment_uploader.rb
@@ -1,19 +1,12 @@
 # frozen_string_literal: true
 
 # Files uploaded as part of a form526 submission that will be sent to EVSS upon form submission.
-class SupportingEvidenceAttachmentUploader < CarrierWave::Uploader::Base
-  include SetAWSConfig
-  include ValidatePdf
-  include ValidateEVSSFileSize
-
-  def size_range
-    1.byte...150.megabytes
-  end
-
+class SupportingEvidenceAttachmentUploader < EVSSClaimDocumentUploaderBase
   def initialize(guid)
     super
     @guid = guid
 
+    #  defaults to CarrierWave::Storage::File if not AWS
     if Rails.env.production?
       set_aws_config(
         Settings.evss.s3.aws_access_key_id,
@@ -22,10 +15,6 @@ class SupportingEvidenceAttachmentUploader < CarrierWave::Uploader::Base
         Settings.evss.s3.bucket
       )
     end
-  end
-
-  def extension_allowlist
-    %w[pdf png gif tiff tif jpeg jpg bmp txt]
   end
 
   def store_dir
